{"schema":"decodex/commit/1","summary":"Split orchestrator status ghost-lane evidence","authority":"XY-1115"}
